Add wrapper function for creating all widgets.
Add separate module for launchin I/O widgets on-the-fly.
