+++ napari_btrack/track.py
@@ -54,42 +54,12 @@ def run_tracker(
         return data, properties, graph
 
 
-def _create_widgets():
-    """Create all the widgets for the plugin"""
-
-    input_widgets = napari_btrack.widgets.create_input_widgets()
-    update_method_widgets = napari_btrack.widgets.create_update_method_widgets()
-    motion_model_widgets = napari_btrack.widgets.create_motion_model_widgets()
-    hypothesis_model_widgets = napari_btrack.widgets.create_hypothesis_model_widgets()
-    control_buttons = napari_btrack.widgets.create_control_widgets()
-
-    widgets = [
-        *input_widgets,
-        *update_method_widgets,
-        *motion_model_widgets,
-        *hypothesis_model_widgets,
-        *control_buttons,
-    ]
-
-    return widgets
-
-
-def _create_default_configs():
-    """Create a set of default configurations for the plugin"""
-
-    # TrackerConfigs automatically loads default cell and particle configs
-    configs = napari_btrack.config.TrackerConfigs()
-    configs["cell"]
-
-    return configs
-
-
 def track() -> Container:
     """Create widgets for the btrack plugin."""
 
     # First create our UI along with some default configs for the widgets
-    all_configs = _create_default_configs()
-    widgets = _create_widgets()
+    widgets = napari_btrack.widgets.create_widgets()
+    all_configs = napari_btrack.config.create_default_configs()
 
     btrack_widget = magicgui.widgets.Container(widgets=widgets, scrollable=True)
     btrack_widget.viewer = napari.current_viewer()
