+++ napari_btrack/widgets/_hypothesis.py
