+++ napari_btrack/widgets/_motion.py
